working

+++ src/main/kotlin/com/lgaertner/euler/usecase/LargestPrimeFactorUseCase.kt
@@ -6,6 +6,6 @@ import com.lgaertner.euler.math.PrimeFactors
 class LargestPrimeFactorUseCase {
     fun execute(number: Long): Long {
         val largestPrimeFactor = Max(PrimeFactors(number))
-        return largestPrimeFactor.value()
+        return largestPrimeFactor.value().toLong()
     }
 }
